Further refactoring to align components based upon best practices.
