+++ client/src/components/footer.js
@@ -21,7 +21,7 @@ import BottomNavigationAction from '@mui/material/BottomNavigationAction';
 import ContactSupportIcon from '@mui/icons-material/ContactSupport';
 // import MessageIcon from '@mui/icons-material/Message'
 import { Link } from "react-router-dom";
-import './footer.css'
+import './assets/css/footer.css'
 
 const styles = {
   footerStyle: {
